test: share deferred-embed scenarios across deferred_embed tests
Extract three reused inline models into scenarios: document_deferred_metadata
(deferred_embed_struct, deferred_embed_update_reloads),
document_metadata_deferred_notes (deferred_field_inside_embed,
update_embed_by_value_with_deferred_sub_field), and
person_contact_deferred_metadata (the two enum-variant deferred tests).
