+++ crates/toasty-driver-integration-suite/src/scenarios.rs
@@ -6,6 +6,8 @@ pub mod counter_value;
 pub mod deferred_document;
 pub mod deferred_json_document;
 pub mod deferred_optional_document;
+pub mod document_deferred_metadata;
+pub mod document_metadata_deferred_notes;
 pub mod fixed_item_name;
 pub mod fixed_item_name_quantity;
 pub mod has_many_belongs_to;
@@ -16,6 +18,7 @@ pub mod has_many_same_target;
 pub mod has_many_unique_title;
 pub mod has_one_optional_belongs_to;
 pub mod in_list_item;
+pub mod person_contact_deferred_metadata;
 pub mod tagged_item;
 pub mod task_bug_priority;
 pub mod task_with_status;
