+++ crates/toasty-driver-integration-suite/src/scenarios/document_deferred_metadata.rs
@@ -1,0 +1,26 @@
+use crate::prelude::*;
+
+scenario! {
+    #![id(ID)]
+
+    #[derive(Debug, toasty::Model)]
+    struct Document {
+        #[key]
+        #[auto]
+        id: ID,
+
+        title: String,
+
+        metadata: toasty::Deferred<Metadata>,
+    }
+
+    #[derive(Debug, toasty::Embed)]
+    struct Metadata {
+        author: String,
+        notes: String,
+    }
+
+    async fn setup(test: &mut Test) -> toasty::Db {
+        test.setup_db(models!(Document)).await
+    }
+}
